// src/Interfaces/Contact/IAddContact.tsx
interface IAddContact {
    statiqueField: {
        userId: number
        languageId: number
        fullName: string
        country: string
        city: string
        email: string
        linkedin: string
        phone: string
        address: string
        postalcode: string
        position: string
        accountId: number
    },
    customFields: []

}
export type { IAddContact }